Add public_token_exchange API
This commit adds `public_token_exchange/2` to exchange a public token for an
access token.

Since it is a public request, it doesn't require an access token, but it does
require credentials, so we add new `credentials` and `authenticated_post`
functions in order to skip adding `access_token` to the request payload.

API Docs: https://plaid.com/docs/api/#exchange-token-flow