Allow Plaidex to be configured using Mix.Config
This allows Plaidex to be configured using Mix.Config. Example:

```
config :plaidex, :plaidex_auth,
  plaid_client_id: "123abc",
  plaid_secret: "123abc"
```

The Config now returns a map instead of a list, which is what the API client
expects. With this commit, it'll successfully make the API calls. Without this
change they all fail because of missing client_id and secret.